refactor(api): 删除根级兼容层与旧顶层导出
Why:
- 在消费者已切到正式入口后，根级 shim 和旧顶层导出已经只剩历史包袱
- 让 `lol_audio_unpack` 根包只保留真正需要的入口，避免继续充当隐形兼容层

What:
- 删除 `app_context.py` 与 `facade.py` 两个根级兼容模块
- 删除 `LolAudioUnpackApp` 的旧 workflow 兼容方法
- 将根包 `__init__.py` 瘦身为仅保留 `__version__` 与 `setup_app`，并让 `setup_app` 走私有内部引用
- 同步更新剩余依赖根级兼容层的测试到 canonical 入口

Risk:
- 兼容性影响：有，根级 shim 和旧顶层导出不再可用
- 潜在风险：如果仓库外仍直接依赖这些入口，升级时需要同步改 import

Test:
- `uv run ruff check src/lol_audio_unpack/app src/lol_audio_unpack/__init__.py src/lol_audio_unpack/runtime/wav tests/test_app_context.py tests/test_app_package_api.py tests/test_remote_preparer.py tests/test_cli.py tests/test_manager_data_updater.py tests/test_manager_utils.py tests/test_wav_transcode.py tests/test_context_injection.py`
- `uv run pytest -q tests/test_app_context.py tests/test_app_package_api.py tests/test_remote_preparer.py tests/test_cli.py tests/test_manager_data_updater.py tests/test_manager_utils.py tests/test_wav_transcode.py tests/test_context_injection.py`
