--- src/lol_audio_unpack/app_context.py
@@ -1,31 +1,0 @@
-"""应用上下文兼容层。
-
-当前首选入口已迁移到 ``lol_audio_unpack.app.context``。
-本模块仅保留旧导入路径兼容，后续阶段会删除。
-"""
-
-from __future__ import annotations
-
-from .app.context import (
-    AppConfig,
-    AppContext,
-    AppContextValidationError,
-    AppPaths,
-    OperationOptions,
-    RemoteSnapshotConfig,
-    SourceMode,
-    WavOutputOptions,
-    create_app_context,
-)
-
-__all__ = [
-    "AppConfig",
-    "AppContext",
-    "AppContextValidationError",
-    "AppPaths",
-    "OperationOptions",
-    "RemoteSnapshotConfig",
-    "SourceMode",
-    "WavOutputOptions",
-    "create_app_context",
-]
